+++ src/mdContainers/mdContainers.hpp
@@ -20,8 +20,8 @@ class mdarray
     private:
         //std::unique_ptr<T[]> Ptr =nullptr;
         T* Ptr=nullptr;
-        std::array<size_t, dim> Size{0};
-        size_t TotalSize=0;
+        std::array<int, dim> Size{0};
+        int TotalSize=0;
         std::array<int, dim> Offset{0};
         MultiIndex<dim> multindex;
         bool NotDestruct = false;
@@ -35,11 +35,11 @@ class mdarray
         mdarray(mdarray<T,dim>&& ToBeMoved);
         mdarray<T,dim>& operator=(mdarray<T,dim>&& ToBeMoved);
 
-        mdarray(const std::array<size_t,dim>& Size_);        
-        void initialize(const std::array<size_t,dim>& Size_);
+        mdarray(const std::array<int,dim>& Size_);        
+        void initialize(const std::array<int,dim>& Size_);
 
-        mdarray(T* Ptr_, const std::array<size_t,dim>& Size_);
-        void initialize(T* Ptr_, const std::array<size_t,dim>& Size_);
+        mdarray(T* Ptr_, const std::array<int,dim>& Size_);
+        void initialize(T* Ptr_, const std::array<int,dim>& Size_);
         
         void fill(const T& FillingValue);
 
@@ -66,7 +66,7 @@ class mdarray
             friend bool operator== (const Iterator& a, const Iterator& b) { return a.m_ptr == b.m_ptr; };
             friend bool operator!= (const Iterator& a, const Iterator& b) { return a.m_ptr != b.m_ptr; };     
             //difference
-            size_t operator-(const Iterator& it2) {return this->m_ptr - it2.m_ptr;}        
+            int operator-(const Iterator& it2) {return this->m_ptr - it2.m_ptr;}        
         private:
             pointer m_ptr;
         };
@@ -91,7 +91,7 @@ class mdarray
         inline T& operator[](const int& oneDindex);        
 
 
-        inline const size_t get_Size(const int& index) const;
+        inline const int get_Size(const int& index) const;
         inline auto get_Size() const {return Size;};
         inline auto get_TotalSize() const {return TotalSize;};
         ~mdarray();
