+++ src/mdContainers/mdContainers_definitions.hpp
@@ -54,7 +54,7 @@ mdarray<T,dim>& mdarray<T,dim>::operator=(mdarray<T,dim>&& ToBeMoved)
 
 
 template<typename T, size_t dim> 
-mdarray<T,dim>::mdarray(const std::array<size_t,dim>& Size_) 
+mdarray<T,dim>::mdarray(const std::array<int,dim>& Size_) 
 {
     this->initialize(Size_);
     this->multindex.initialize(Size_);
@@ -62,7 +62,7 @@ mdarray<T,dim>::mdarray(const std::array<size_t,dim>& Size_)
 }
 
 template<typename T, size_t dim> 
-mdarray<T,dim>::mdarray(T* Ptr_, const std::array<size_t,dim>& Size_)
+mdarray<T,dim>::mdarray(T* Ptr_, const std::array<int,dim>& Size_)
 {
     this->initialize(Ptr_, Size_);
     this->multindex.initialize(Size_);
@@ -82,7 +82,7 @@ void mdarray<T,dim>::TotalSizeAndOffset()
 }
 
 template<typename T, size_t dim> 
-void mdarray<T,dim>::initialize(const std::array<size_t,dim>& Size_)
+void mdarray<T,dim>::initialize(const std::array<int,dim>& Size_)
 {
     (*this).~mdarray<T,dim>();
     Size = Size_;
@@ -96,7 +96,7 @@ void mdarray<T,dim>::initialize(const std::array<size_t,dim>& Size_)
 }
 
 template<typename T, size_t dim> 
-void mdarray<T,dim>::initialize(T* Ptr_, const std::array<size_t,dim>& Size_)
+void mdarray<T,dim>::initialize(T* Ptr_, const std::array<int,dim>& Size_)
 {
     Ptr = Ptr_; 
     Size = Size_; 
@@ -158,7 +158,7 @@ inline T& mdarray<T,dim>::operator[](const int& oneDindex)
 
 
 template <typename T, size_t dim>
-inline const size_t mdarray<T,dim>::get_Size(const int& index) const
+inline const int mdarray<T,dim>::get_Size(const int& index) const
 {
     return Size[index];
 }
